Update to fix two implementation errors
First error: the optimization optimized to the SP_* values.  This was fixed by the changed math in the return statement in HoleyCalibration.LengthDeltaFromIdeal.
Second error: the kinematics 'D' property updates appeared to have no affect.  This was fixed by exercising the kinematics.recomputeGeometry() method.
Deleted unneeded 'LengthChangePlusOne' method, which was an unsuccessful attempt to solve an earlier issue.